+++ weblate/accounts/views.py
@@ -792,11 +792,19 @@ def user_avatar(request: AuthenticatedHttpRequest, user: str, size: int):
         raise Http404(msg)
 
     avatar_user = get_object_or_404(User, username=user)
+    email = avatar_user.email
 
-    if avatar_user.email == "noreply@weblate.org":
-        return redirect(get_fallback_avatar_url(int(size)))
-    if avatar_user.email == f"noreply+{avatar_user.pk}@weblate.org":
-        return redirect(os.path.join(settings.STATIC_URL, "state/ghost.svg"))
+    # Deleted users
+    if email == f"noreply+{avatar_user.pk}@weblate.org":
+        return redirect(
+            os.path.join(settings.STATIC_URL, "state/ghost.svg"), permanent=True
+        )
+    # Bot and anonymous accounts
+    if email.startswith("noreply") and email.endswith("@weblate.org"):
+        return redirect(get_fallback_avatar_url(int(size)), permanent=True)
+    # Project API tokens
+    if email.endswith("@bots.noreply.weblate.org"):
+        return redirect(get_fallback_avatar_url(int(size), "api"), permanent=True)
 
     response = HttpResponse(
         content_type="image/png", content=get_avatar_image(avatar_user, size)
@@ -853,7 +861,7 @@ class WeblateLoginView(BaseLoginView):
         context = super().get_context_data(**kwargs)
         auth_backends = get_auth_keys()
         context["login_backends"] = [x for x in sorted(auth_backends) if x != "email"]
-        context["can_reset"] = "email" in auth_backends
+        context["login_email"] = context["can_reset"] = "email" in auth_backends
         context["title"] = gettext("Sign in")
         return context
 
@@ -867,6 +875,9 @@ class WeblateLoginView(BaseLoginView):
         auth_backends = get_auth_keys()
         if len(auth_backends) == 1 and "email" not in auth_backends:
             return redirect_single(request, auth_backends.pop())
+
+        if request.method == "POST" and "email" not in auth_backends:
+            return redirect("login")
 
         return super().dispatch(request, *args, **kwargs)