+++ weblate/api/views.py
@@ -625,10 +625,12 @@ class GroupViewSet(viewsets.ModelViewSet):
 
     def get_queryset(self):
         if self.request.user.has_perm("group.edit"):
-            return Group.objects.order_by("id")
-        return self.request.user.groups.order_by(
-            "id"
-        ) | self.request.user.administered_group_set.order_by("id")
+            queryset = Group.objects.all()
+        else:
+            queryset = Group.objects.filter(
+                Q(user=self.request.user) | Q(admins=self.request.user)
+            ).distinct()
+        return queryset.order_by("id")
 
     def perm_check(
         self,
@@ -1427,10 +1429,9 @@ class ComponentViewSet(
             if not obj.can_add_new_language(request.user):
                 self.permission_denied(request, message=obj.new_lang_error_message)
 
-            if request.user.has_perm("translation.add_more", obj):
-                base_languages = obj.get_all_available_languages()
-            else:
-                base_languages = obj.get_available_languages()
+            base_languages = obj.get_all_available_languages()
+            if not request.user.has_perm("translation.add_more", obj):
+                base_languages = base_languages.filter_for_add(obj.project)
 
             try:
                 language = base_languages.get(code=language_code)