+++ weblate/examples/fix_foo.py
@@ -1,7 +1,3 @@
-# Copyright © Michal Čihař <michal@weblate.org>
-#
-# SPDX-License-Identifier: GPL-3.0-or-later
-
 from django.utils.translation import gettext_lazy
 
 from weblate.trans.autofixes.base import AutoFix