+++ weblate/memory/tests.py
@@ -23,6 +23,7 @@ from weblate.memory.utils import CATEGORY_FILE
 from weblate.trans.tests.test_views import FixtureTestCase
 from weblate.trans.tests.utils import get_test_file
 from weblate.utils.db import TransactionsTestMixin
+from weblate.utils.hash import hash_to_checksum
 from weblate.utils.state import STATE_TRANSLATED
 
 
@@ -35,6 +36,7 @@ def add_document() -> None:
         origin="test",
         from_file=True,
         shared=False,
+        status=Memory.STATUS_ACTIVE,
     )
 
 
@@ -117,6 +119,8 @@ class MemoryModelTest(TransactionsTestMixin, FixtureTestCase):
                     "target": "Ahoj",
                     "origin": "test",
                     "category": CATEGORY_FILE,
+                    "status": 1,
+                    "context": "",
                 }
             ],
         )
@@ -264,6 +268,234 @@ class MemoryModelTest(TransactionsTestMixin, FixtureTestCase):
         self.assertEqual(Memory.objects.count(), 3)
         handle_unit_translation_change(unit, self.user)
         self.assertEqual(Memory.objects.count(), 3)
+
+    def test_memory_status_no_review(self) -> None:
+        self.test_memory_status_with_review(translation_review=False)
+
+    def test_memory_status_with_review(self, translation_review: bool = True) -> None:
+        self.project.translation_review = translation_review
+        self.project.save()
+        machine_translation = WeblateMemory({})
+
+        unit = self.get_unit()
+        unit.translate(self.user, "Hello", STATE_TRANSLATED)
+
+        # check memory status is created with status pending
+        expected_status = (
+            Memory.STATUS_PENDING if translation_review else Memory.STATUS_ACTIVE
+        )
+        self.assertEqual(
+            1,
+            Memory.objects.filter(project=self.project, status=expected_status).count(),
+        )
+
+        suggestion = machine_translation.search(unit, "Hello, world!\n", None)[0]
+        if translation_review:
+            # quality is less than 100% because of penalty
+            self.assertLess(suggestion["quality"], 100)
+        else:
+            self.assertEqual(suggestion["quality"], 100)
+
+        if translation_review:
+            self.approve_translation(unit, "Hello")
+
+            # check that memory status is updated to active
+            self.assertEqual(
+                1,
+                Memory.objects.filter(
+                    project=self.project, status=Memory.STATUS_ACTIVE
+                ).count(),
+            )
+            suggestion = machine_translation.search(unit, "Hello, world!\n", None)[0]
+            self.assertEqual(suggestion["quality"], 100)
+
+            # mark the translation as needing editing
+            self.approve_translation(unit, "Hello", review="10")
+
+            # check that memory status is updated to pending
+            self.assertEqual(
+                1,
+                Memory.objects.filter(
+                    project=self.project, status=Memory.STATUS_PENDING
+                ).count(),
+            )
+            suggestion = machine_translation.search(unit, "Hello, world!\n", None)[0]
+            self.assertLess(suggestion["quality"], 100)
+
+    def approve_translation(self, unit, target: str, review: str = "30"):
+        # allow user to approve translations
+        self.project.add_user(self.user, "Administration")
+        params = {
+            "checksum": unit.checksum,
+            "contentsum": hash_to_checksum(unit.content_hash),
+            "translationsum": hash_to_checksum(unit.get_target_hash()),
+            "target_0": target,
+            "review": review,
+        }
+        self.client.post(unit.translation.get_translate_url(), params, follow=True)
+
+    def test_pending_memory_autoclean(self, autoclean_active: bool = False) -> None:
+        import_memory(self.project.id)
+        imported_memory_ids = [m.pk for m in Memory.objects.all()]
+        initial_memory_count = len(imported_memory_ids)
+        not_imported_memory_qs = Memory.objects.exclude(id__in=imported_memory_ids)
+
+        # create a translation with review enabled
+        self.project.translation_review = True
+        self.project.save()
+        machine_translation = WeblateMemory({})
+
+        unit = self.get_unit()
+        unit.translate(self.user, "Hello 1", STATE_TRANSLATED)
+
+        # check memory status is created with status pending
+        self.assertEqual(
+            1,
+            not_imported_memory_qs.filter(
+                project=self.project, context=unit.context, status=Memory.STATUS_PENDING
+            ).count(),
+        )
+
+        # check that suggestion quality is less than 100% because of penalty
+        suggestion = machine_translation.search(unit, "Hello, world!\n", None)[0]
+        self.assertLess(suggestion["quality"], 100)
+
+        # another user submits a translation
+        unit.translate(self.anotheruser, "Hello 2", STATE_TRANSLATED)
+        self.assertEqual(
+            2,
+            not_imported_memory_qs.filter(
+                project=self.project, context=unit.context, status=Memory.STATUS_PENDING
+            ).count(),
+        )
+        for suggestion in machine_translation.search(unit, "Hello, world!\n", None):
+            self.assertLess(suggestion["quality"], 100)
+
+        # approve one translation, check that only 1 memory left with status active
+        self.approve_translation(unit, "Hello 1")
+
+        self.assertEqual(
+            1,
+            not_imported_memory_qs.filter(
+                project=self.project, context=unit.context, status=Memory.STATUS_ACTIVE
+            ).count(),
+        )
+        suggestion = machine_translation.search(unit, "Hello, world!\n", None)[0]
+        self.assertEqual(suggestion["quality"], 100)
+
+        if not autoclean_active:
+            # check that the other pending memory has not been deleted
+            self.assertEqual(
+                1,
+                not_imported_memory_qs.filter(
+                    project=self.project,
+                    context=unit.context,
+                    status=Memory.STATUS_PENDING,
+                ).count(),
+            )
+            for suggestion in machine_translation.search(unit, "Hello, world!\n", None):
+                if suggestion["text"] == "Hello 2\n":  # ignore imported entries
+                    self.assertLess(suggestion["quality"], 100)
+
+        # check that imported memory entries were not affected by autoclean
+        self.assertEqual(
+            initial_memory_count,
+            Memory.objects.filter(pk__in=imported_memory_ids).count(),
+        )
+
+    def test_pending_memory_autoclean_active(self) -> None:
+        self.project.autoclean_tm = True
+        self.project.save()
+        self.test_pending_memory_autoclean(autoclean_active=True)
+
+    def test_clean_memory_command(
+        self, autoclean_tm: bool = False, translation_review: bool = False
+    ) -> None:
+        self.project.autoclean_tm = autoclean_tm
+        self.project.save()
+
+        import_memory(self.project.id)
+        excepted_deleted_count = 0
+
+        unit = self.get_unit()
+
+        self.project.translation_review = translation_review
+        self.project.save()
+
+        unit.translate(self.user, "Hello 1", STATE_TRANSLATED)
+        unit.translate(self.anotheruser, "Hello 2", STATE_TRANSLATED)
+
+        if translation_review:
+            self.approve_translation(unit, "Hello 1")
+            if not autoclean_tm:
+                excepted_deleted_count += 3  # 1 for each [project, user, shared]
+
+        total_memory_count = Memory.objects.count()
+        call_command("cleanup_memory")
+        self.assertEqual(
+            Memory.objects.all().count(), total_memory_count - excepted_deleted_count
+        )
+
+    def test_clean_memory_command_with_translation_review_no_autoclean(self) -> None:
+        self.test_clean_memory_command(autoclean_tm=False, translation_review=True)
+
+    def test_clean_memory_command_with_translation_review_and_autoclean(self) -> None:
+        self.test_clean_memory_command(autoclean_tm=True, translation_review=True)
+
+    def test_clean_memory_command_no_translation_review_with_autoclean(self) -> None:
+        self.test_clean_memory_command(autoclean_tm=True, translation_review=False)
+
+    def test_memory_context(
+        self, autoclean_tm: bool = False, translation_review: bool = False
+    ) -> None:
+        self.project.translation_review = translation_review
+        self.project.autoclean_tm = autoclean_tm
+        self.project.save()
+        machine_translation = WeblateMemory({})
+
+        # check that quality of suggestion is inferior if contexts differ
+        unit = self.get_unit()
+        unit2 = self.translation.unit_set.create(
+            context="Different context",
+            source=unit.source,
+            source_unit=unit.source_unit,
+            id_hash=1001,
+            position=1001,
+        )
+        unit.translate(self.user, "Hello no context", STATE_TRANSLATED)
+        unit2.translate(self.user, "Hello with context", STATE_TRANSLATED)
+
+        if translation_review:
+            self.approve_translation(unit, "Hello no context")
+            self.approve_translation(unit2, "Hello with context")
+
+        suggestions = machine_translation.search(unit, unit.source, None)
+        with_context = [s for s in suggestions if "Hello with context" in s["text"]][0]  # noqa: RUF015
+        no_context = [s for s in suggestions if "Hello no context" in s["text"]][0]  # noqa: RUF015
+        self.assertLess(with_context["quality"], no_context["quality"])
+
+        # check that memory with different context is not affected by autoclean
+        if autoclean_tm:
+            unit.translate(self.user, "New translation", STATE_TRANSLATED)
+            self.approve_translation(unit, "New translation")
+            suggestions = machine_translation.search(unit, unit.source, None)
+
+            self.assertFalse(
+                [s for s in suggestions if "Hello no context" in s["text"]]
+            )
+            self.assertTrue([s for s in suggestions if "New translation" in s["text"]])
+            self.assertTrue(
+                [s for s in suggestions if "Hello with context" in s["text"]]
+            )
+
+    def test_memory_context_with_review_no_autoclean(self):
+        self.test_memory_context(False, True)
+
+    def test_memory_context_with_review_and_autoclean(self):
+        self.test_memory_context(True, True)
+
+    def test_memory_context_no_review_with_autoclean(self):
+        self.test_memory_context(True, False)
 
 
 class MemoryViewTest(FixtureTestCase):