+++ weblate/trans/management/commands/benchmark.py
@@ -2,12 +2,12 @@
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-import cProfile
-import os
-import pstats
+
+from django.test.utils import override_settings
 
 from weblate.trans.models import Component, Project
 from weblate.utils.management.base import BaseCommand
+from weblate.utils.views import create_component_from_zip
 
 
 class Command(BaseCommand):
@@ -17,61 +17,35 @@ class Command(BaseCommand):
 
     def add_arguments(self, parser) -> None:
         super().add_arguments(parser)
-        prefix = os.path.dirname(
-            os.path.dirname(os.path.dirname(os.path.dirname(__file__)))
-        )
-        parser.add_argument(
-            "--profile-sort", default="cumulative", help="sort order for profile stats"
-        )
-        parser.add_argument(
-            "--profile-filter",
-            default=prefix,
-            help=f"filter for profile stats, defaults to {prefix}",
-        )
-        parser.add_argument(
-            "--profile-count",
-            type=int,
-            default=20,
-            help="number of profile stats to show",
-        )
         parser.add_argument("--template", default="", help="template monolingual files")
-        parser.add_argument(
-            "--delete", action="store_true", help="delete after testing"
-        )
+        parser.add_argument("--keep", action="store_true", help="keep after testing")
         parser.add_argument("--format", default="po", help="file format")
-        parser.add_argument("project", help="Existing project slug for tests")
-        parser.add_argument("repo", help="Test VCS repository URL")
-        parser.add_argument("mask", help="File mask")
+        parser.add_argument("--project", help="Existing project slug for tests")
+        parser.add_argument("--repo", help="Test VCS repository URL")
+        parser.add_argument("--filemask", help="File mask")
+        parser.add_argument("--zipfile", help="Zip file")
 
     def handle(self, *args, **options) -> None:
-        project = Project.objects.get(slug=options["project"])
-        # Delete any possible previous tests
-        Component.objects.filter(project=project, slug="benchmark").delete()
-        profiler = cProfile.Profile()
-        component = profiler.runcall(
-            Component.objects.create,
-            name="Benchmark",
-            slug="benchmark",
-            repo=options["repo"],
-            filemask=options["mask"],
-            template=options["template"],
-            file_format=options["format"],
-            project=project,
-        )
-        profiler.enable()
-        component.after_save(
-            changed_git=True,
-            changed_setup=False,
-            changed_template=False,
-            changed_variant=False,
-            changed_enforced_checks=False,
-            skip_push=True,
-            create=True,
-        )
-        profiler.disable()
-        stats = pstats.Stats(profiler, stream=self.stdout)
-        stats.sort_stats(options["profile_sort"])
-        stats.print_stats(options["profile_filter"], options["profile_count"])
-        # Delete after testing
-        if options["delete"]:
-            component.delete()
+        # Execute tasks in place
+        with override_settings(CELERY_TASK_ALWAYS_EAGER=True):
+            project = Project.objects.get(slug=options["project"])
+            # Delete any possible previous tests
+            Component.objects.filter(project=project, slug="benchmark").delete()
+            params = {
+                "name": "Benchmark",
+                "slug": "benchmark",
+                "repo": options["repo"],
+                "filemask": options["filemask"],
+                "template": options["template"],
+                "file_format": options["format"],
+                "project": project,
+            }
+            if options["zipfile"]:
+                params["vcs"] = "local"
+                params["repo"] = "local:"
+                params["branch"] = "main"
+                create_component_from_zip(params, options["zipfile"])
+            component = Component.objects.create(**params)
+            # Delete after testing
+            if not options["keep"]:
+                component.delete()