+++ weblate/trans/models/translation.py
@@ -585,6 +585,11 @@ class Translation(
         if self.is_source:
             self.component.preload_sources(updated)
 
+        # Unload the store, this is intentionally in a way that it would break
+        # further consumers as no further consumer is expected after this and
+        # we do not want to parse the file again.
+        self.__dict__["store"] = None
+
     def store_update_changes(self) -> None:
         # Save change
         Change.objects.bulk_create(self.update_changes, batch_size=500)
@@ -751,7 +756,7 @@ class Translation(
         grouped according to the following logic:
 
         1. Maintain active groups per author in a dict ({author_id: {"changes": [...], "units": {...}}})
-        2. For each pending change, check if the unit conflicts with ANY active group
+        2. For each pending change, check if the unit conflicts with ANY active group EXCEPT the change author's own group.
         3. If conflict found: finalize conflicting group(s)
         4. Add change to the author's existing group or create a new one if needed
 
@@ -762,12 +767,13 @@ class Translation(
             - A2: Author 2 edits Unit A -> Unit A conflicts -> Finalize Group 1, start Group 2: Author2
             - B3: Author 1 edits Unit B -> No conflicts -> Start Group 3: Author1
             - C4: Author 2 edits Unit C -> No conflicts -> Add to Group 2: Author2
-            - D5: Author 1 edits Unit D -> No conflicts -> Add to Group 3: Author1
+            - B5: Author 1 edits Unit B -> conflicts with own change -> Do not break group, Add to Group 3: Author1
+            - D6: Author 1 edits Unit D -> No conflicts -> Add to Group 3: Author1
 
         Result:
             - Commit 1: Author1 [A1]
             - Commit 2: Author2 [A2, C4]
-            - Commit 3: Author1 [B3, D5]
+            - Commit 3: Author1 [B3, B5, D6]
 
         """
         author_map = {
@@ -785,7 +791,7 @@ class Translation(
 
             conflicting_authors = []
             for other_author_id, group_data in active_groups.items():
-                if unit_id in group_data["units"]:
+                if unit_id in group_data["units"] and author_id != other_author_id:
                     conflicting_authors.append(other_author_id)
 
             for conflicting_author_id in conflicting_authors: